Cache result of last attribute matching
In case of multiple candidates during attribute matching, it's very
frequent that we ask for disambiguation of the same set of attributes.
In practice, the candidates are different (because they correspond to
different variants of different modules), but their attributes are the
same (they all have the number of variants, with the same attributes).
So we can cache the result of disambiguation, and "map" the result
back to the candidate list.